// public/core/components/menu/types.d.ts
export type TMenuItemTrigger = Record<string, TMenuToggle> | TMenuTrigger;
export type TMenuToggle = 'accordion' | 'dropdown';
export type TMenuItemToggle = Record<string, TMenuToggle> | TMenuToggle;
export type TMenuDropdown = Partial<Omit<PopperProps, 'children'>>;
// ...
export interface IMenuItemConfig {
    title?: string;
    disabled?: boolean;
    heading?: string;
    icon?: string;
    badge?: string;
    separator?: boolean;
    tooltip?: Partial<TooltipProps>;
    path?: string;
    rootPath?: string;
    bullet?: boolean;
    collapse?: boolean;
    collapseTitle?: string;
    expandTitle?: string;
    toggle?: TMenuItemToggle;
    dropdownProps?: TMenuDropdown;
    trigger?: TMenuItemTrigger;
    children?: IMenuItemConfig[];
    childrenIndex?: number;
}
export type TMenuConfig = IMenuItemConfig[];